requirements.txt
Flask-Markdown==3.3.4
spacy==2.2.4
awscli==1.20.27
botocore==1.21.27
s3cmd==2.1.0